reflect that state is about transactions in naming
