+++ deployables/extension/src/panel/state/isConfirmedTransaction.ts
@@ -1,18 +1,4 @@
-import type { MetaTransactionRequest } from '@zodiac/schema'
-import type { ContractInfo } from '../utils'
-
-export type UnconfirmedTransaction = MetaTransactionRequest & {
-  id: string
-  createdAt: Date
-  contractInfo?: ContractInfo
-}
-
-export type ConfirmedTransaction = UnconfirmedTransaction & {
-  snapshotId: string
-  transactionHash: string
-}
-
-export type Transaction = UnconfirmedTransaction | ConfirmedTransaction
+import type { ConfirmedTransaction, Transaction } from './state'
 
 export const isConfirmedTransaction = (
   transaction: Transaction,
