+++ deployables/extension/test-utils/RenderWrapper.tsx
@@ -1,5 +1,5 @@
 import { ProvidePort } from '@/port-handling'
-import { ProvideState, type State } from '@/state'
+import { ProvideTransactions, type State } from '@/state'
 import { type PropsWithChildren } from 'react'
 
 type RenderWraperProps = PropsWithChildren<{
@@ -10,9 +10,9 @@ export const RenderWrapper = ({
   children,
   initialState,
 }: RenderWraperProps) => (
-  // <ProvideCompanionAppContext url={getCompanionAppUrl()}>
   <ProvidePort>
-    <ProvideState initialState={initialState}>{children}</ProvideState>
+    <ProvideTransactions initialState={initialState}>
+      {children}
+    </ProvideTransactions>
   </ProvidePort>
-  // </ProvideCompanionAppContext>
 )
